Move format to bot src/util
