+++ matrix-meetings-bot/src/util/helpers.ts
